Update integrations.py
Major changes to audit function and logging autogenerated values, such as PPIDs and visit labels -- specimen label logging may be added in the future, but is more complex and not as pressing, since it is unlikely that specimens will be bulk created in this fashion without already having labels (i.e. as migration from another system/DB). It is generally true that participants and visits will already have labels in cases of migrating data, but we have seen many more instances of PPIDs and visit labels containing PHI that needs to be stripped out/stored in a different field, than we have for specimen labels. It is also the case that specimen labels are often physical things that need to be replaced on vials, blocks, slides, etc. Consequentially, we have seen hesitance from our users specifically when addressing specimen labeling.